Mod compatibility check is now in a separate class and not part of ModHandler
